+++ app/Http/Requests/UpdateSettings.php
@@ -65,6 +65,7 @@ class UpdateSettings extends FormRequest
             'room_auto_delete_inactive_period' => ['required', 'numeric', Rule::enum(TimePeriod::class)],
             'room_auto_delete_never_used_period' => ['required', 'numeric', Rule::enum(TimePeriod::class)],
             'room_auto_delete_deadline_period' => ['required', 'numeric', Rule::enum(TimePeriod::class)->only([TimePeriod::ONE_WEEK, TimePeriod::TWO_WEEKS, TimePeriod::ONE_MONTH])],
+            'room_file_terms_of_use' => ['nullable', 'string', 'max:65000'],
 
             'user_password_change_allowed' => ['required', 'boolean'],